django-appointments/requirements.txt
anyjson==0.3.3
django-crispy-forms==1.9.2

